Remove `prefer` function.
Remove use of the defunct preferred concept that is no longer available
in Paxos. It was a way to indicate that we would prefer certain citizens
to be legislators. This can be as easily done externally by submitting
requests to change the shape of the government, by promotion one citizen
at a time up one run at a time, as opposed to internally with a more
convoluted election logic.