Remove multi-part synchronization.
Remove the notion of a synchronization footer. It used to be the case
that synchronization would synchronize user state, so things where
already complicated, and the footer was there because my thinking about
the matter had become complicated.

Now Kibitz is going to only be a matter of wrapping Paxos, so it will be
consensus, a log, and a government and the user can make use of these
properties, building a replicated state machine, if that is what they
want, and a key value store on top of that, if that is what they want.