Create a single end point for messaging.
This makes it simpler to implement in an existing server, or to wrap a
server around it quickly, only one end point and you hand the messages
off to the Kibitzer. It would be simpler to implement, but not very
restful. This isn't an API however, no one else is supposed to use these
end points.

This seems like you have an intuition and the only thing holding you
back is pedantry. There's something about designing an API for this that
invites unwanted consideration from the user. In this case, HTTP is just
a transport that makes some guarantees. It is below the application in
the network stack.