Break the pump from Paxos to Islander.
Break the pump from Paxos to Islander. Currently Compassion Colleague is
pushing a null through the `kibitzer.paxos.log` to explicitly end the
stream but Paxos checks the top of `paxos.log` often, using it as a most
recent log entry variable. When an entry pushed onto the `paxos.log` in
destroys the Compassion Colleague, when the log push returns the top of
the `paxos.log` is null and Paxos cannot complete it's handling of the
new entry event.

Breaking this using the shifter destructor is the right way to do it.